Изменил тип данных для хранения размера шкалы

+++ TimeDiagrammGeneratorLibrary/GraphicObjects/IScale.cs
@@ -3,6 +3,6 @@
     public interface IScale
     {
         int Size { get; }
-        float DotsPerDivision { get; }
+        int DotsPerDivision { get; }
     }
 }
